Zebra heatmap WIP
